==> 1_preprocess/extendAtlas/1.8_permute_cpm2.ipynb <==
import scanpy as sc
import numpy as np
import pandas as pd
import decoupler
import seaborn
from itertools import compress

import sys
sys.path.append("../../function/")
import trajDiff_v133 as trajDiff
milo2=trajDiff.Milo2()

extendAtlas=sc.read("../../important_processed_data/9.18_intergrate.h5ad")

extendMeta=pd.read_csv("../../important_processed_data/24.1.7_extendAtlasObs.csv",index_col=0)

extendAtlas.obs=extendMeta

extendAtlas

fibroLineage=extendAtlas[extendAtlas.obs["lineage_laFibro"]].copy()

fibroLineage

mdata_lineage=milo2.load(fibroLineage)

sc.pp.neighbors(mdata_lineage['rna'], use_rep='scANVI', n_neighbors=200, n_pcs=15)
milo2.make_nhoods(mdata_lineage['rna'], prop=0.015)
mdata_lineage = milo2.count_nhoods(mdata_lineage['rna'], sample_col="Sample")
pseudobulk_lineage=milo2.make_pseudobulk_parallel(mdata=mdata_lineage,sample_col="Sample",group_col="Stage",time_col="dpt_pred",njob=25,min_cell=2)
milo2.make_range(mdata_lineage,only_range=True)


wholeCpm=milo2.make_whole_cpm(mdata_lineage,n_jobs=25)
milo2.permute_point_cpm_parallel(mdata_lineage,n_jobs=25)


cpmDict_fibro=mdata_lineage["milo"].uns["cpm_dict"]

!mkdir ../process_data/traj_diff/cpm/fibro/

# Specify the directory where you want to save the CSV files
output_directory = '../process_data/traj_diff/cpm/fibro/'

# Export each DataFrame in the dictionary to a separate CSV file
for df_name, df in cpmDict_fibro.items():
    df.to_csv(output_directory + f'{df_name}.csv', index=True)

chondroLineage=extendAtlas[extendAtlas.obs["lineage_chondro"]].copy()

mdata_lineage=milo2.load(chondroLineage)

sc.pp.neighbors(mdata_lineage['rna'], use_rep='scANVI', n_neighbors=200, n_pcs=15)
milo2.make_nhoods(mdata_lineage['rna'], prop=0.015)
mdata_lineage = milo2.count_nhoods(mdata_lineage['rna'], sample_col="Sample")
pseudobulk_lineage=milo2.make_pseudobulk_parallel(mdata=mdata_lineage,sample_col="Sample",group_col="Stage",time_col="dpt_pred",njob=25,min_cell=2)
milo2.make_range(mdata_lineage,only_range=True)

mdata_lineage["milo"].write_h5ad("../process_data/traj_diff/fibro_milo.h5ad")
mdata_lineage["pseudobulk"].write_h5ad("../process_data/traj_diff/fibro_pseudobulk.h5ad")

lineageMilo=sc.read_h5ad("../process_data/traj_diff/fibro_milo.h5ad")
lineagePesudo=sc.read_h5ad("../process_data/traj_diff/fibro_pseudobulk.h5ad")

mdata_lineage=milo2.load(lineageMilo)

mdata_lineage.mod["milo"]=lineageMilo
mdata_lineage.mod["pseudobulk"]=lineagePesudo


wholeCpm=milo2.make_whole_cpm(mdata_lineage,n_jobs=25)
milo2.permute_point_cpm_parallel(mdata_lineage,n_jobs=25)


!mkdir ../process_data/traj_diff/cpm/chondro/

cpmDict=mdata_lineage["milo"].uns["cpm_dict"]
# Specify the directory where you want to save the CSV files
output_directory = '../process_data/traj_diff/cpm/chondro/'

# Export each DataFrame in the dictionary to a separate CSV file
for df_name, df in cpmDict.items():
    df.to_csv(output_directory + f'{df_name}.csv', index=True)

milo2.permute_point_cpm_parallel(mdata_lineage,n_jobs=25)

pseudobulk_lineage=sc.read_h5ad("../process_data/traj_diff/lepr_pseudobulk.h5ad")
milo_lineage=sc.read_h5ad("../process_data/traj_diff/milo_lineage_lepr.h5ad")

mdata_lineage=milo2.load(milo_lineage)
mdata_lineage.mod["milo"]=milo_lineage
mdata_lineage.mod["pseudobulk"]=pseudobulk_lineage


wholeCpm=milo2.make_whole_cpm(mdata_lineage,n_jobs=25)
milo2.permute_point_cpm_parallel(mdata_lineage,n_jobs=25)


!mkdir ../process_data/traj_diff/cpm/lepr/

cpmDict=mdata_lineage["milo"].uns["cpm_dict"]
# Specify the directory where you want to save the CSV files
output_directory = '../process_data/traj_diff/cpm/lepr/'

# Export each DataFrame in the dictionary to a separate CSV file
for df_name, df in cpmDict.items():
    df.to_csv(output_directory + f'{df_name}.csv', index=True)
